+++ frontend/src/shared/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useScreenWidth'
 export * from './useToggle'
+export * from './useFileUpload'
